requirements.txt
langchain==0.0.204
google-generativeai==0.1.0
langchain_google_genai==0.0.2
python-dotenv==1.0.0
rich==13.4.2
httpx==0.24.1